+++ Views/MainWindow.xaml
@@ -1,0 +1,97 @@
+﻿<Window x:Class="Wukong_PBData_ReadWriter_GUI.Views.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:vm="clr-namespace:Wukong_PBData_ReadWriter_GUI.ViewModels"
+        xmlns:hc="https://handyorg.github.io/handycontrol"
+        mc:Ignorable="d"
+        Title="黑猴配表编辑器" Height="720" Width="1280">
+    <Window.DataContext>
+        <vm:MainWindowViewModel />
+    </Window.DataContext>
+    <DockPanel>
+        <!-- 顶部菜单栏 -->
+        <Menu DockPanel.Dock="Top">
+            <MenuItem Header="文件">
+                <MenuItem Header="打开Data目录" Click="OpenDataFolder" />
+                <MenuItem Header="保存Data文件" Click="SaveDataFile" />
+                <MenuItem Header="另存为Data文件" Click="SaveAsNewDataFile" />
+                <Separator />
+                <MenuItem Header="导入备注配置" Click="ImportDescription" />
+                <MenuItem Header="导出备注配置" Click="ExportDescription" />
+                <Separator />
+                <MenuItem Header="生成PAK" Click="CreatePak" />
+                <MenuItem Header="解包PAK" Click="UncompressPak" />
+                <MenuItem Header="合并PAK" Click="CreateUnpackWindow" />
+                <Separator />
+                <MenuItem Header="退出" Click="Close_Click" />
+            </MenuItem>
+            <MenuItem Header="帮助">
+                <MenuItem Header="查看帮助" Click="HelpMenuItem_Click" />
+                <MenuItem Header="关于作者" Click="AuthorMenuItem_Click_1" />
+            </MenuItem>
+        </Menu>
+
+        <StackPanel DockPanel.Dock="Top" Margin="5">
+            <TextBox Name="GlobalSearchBox" Width="Auto" Margin="0,0,0,10" Text="全局搜索" Foreground="Gray"
+                     GotFocus="GlobalSearchBox_GotFocus" LostFocus="GlobalSearchBox_LostFocus"
+                     TextChanged="GlobalSearchBox_TextChanged" />
+            <Expander Name="SearchResultsExpander" Header="搜索结果" IsExpanded="False" Visibility="Collapsed">
+                <ListBox Name="SearchResultsList" Width="Auto" Height="200"
+                         ScrollViewer.VerticalScrollBarVisibility="Auto" />
+            </Expander>
+        </StackPanel>
+
+        <!-- 主体内容 -->
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="2*" />
+                <ColumnDefinition Width="3*" />
+                <ColumnDefinition Width="5*" />
+            </Grid.ColumnDefinitions>
+            <!-- 第一个区块 -->
+            <StackPanel Grid.Column="0" Margin="5">
+                <hc:TextBox Width="Auto" Margin="0,0,0,10" Foreground="Gray"
+                            hc:InfoElement.Placeholder="搜索配表文件"
+                            Text="{Binding FileSearchText,UpdateSourceTrigger=PropertyChanged}" />
+                <ListBox x:Name="TopFileList" Width="Auto" Height="120">
+                </ListBox>
+                <ListBox x:Name="FileList" Width="Auto" Height="460"
+                         ItemsSource="{Binding FilteredDataFiles}"
+                         DisplayMemberPath="FileInfo.Name"
+                         SelectedItem="{Binding SelectedFile,Mode=OneWayToSource}" />
+            </StackPanel>
+
+            <!-- 第二个区块 -->
+            <StackPanel Grid.Column="1" Margin="5">
+                <hc:TextBox Width="Auto" Margin="0,0,0,10" Foreground="Gray"
+                            hc:InfoElement.Placeholder="搜索ID或备注" />
+                <ListBox x:Name="DataItemList" Width="Auto" Height="540"
+                         ItemsSource="{Binding SelectedFile.DataItemList}"
+                         DisplayMemberPath="Desc"
+                         SelectedItem="{Binding SelectedItem,Mode=OneWayToSource}" />
+                <Button Content="新增一项" Width="Auto" Height="40" Click="AddNewDataItem_Click" />
+            </StackPanel>
+
+            <!-- 第三个区块 -->
+            <StackPanel Grid.Column="2" Margin="5">
+                <TextBlock Name="DataFilePath" Text="配置数据" FontWeight="Bold" Margin="0,0,0,10" />
+
+                <ScrollViewer VerticalScrollBarVisibility="Auto" Height="600">
+                    <ItemsControl ItemsSource="{Binding SelectedItem.DataPropertyItems}">
+                        <ItemsControl.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Label Content="{Binding PropertyName}" Width="320" 
+                                           ToolTip="{Binding PropertyDesc}"/>
+                                    <TextBlock Text="{Binding PropertyDesc}" Width="320" />
+                                </StackPanel>
+                            </DataTemplate>
+                        </ItemsControl.ItemTemplate>
+                    </ItemsControl>
+                </ScrollViewer>
+            </StackPanel>
+        </Grid>
+    </DockPanel>
+</Window>
