+++ Views/MergeWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="Wukong_PBData_ReadWriter_GUI.MergeWindow"
+﻿<Window x:Class="Wukong_PBData_ReadWriter_GUI.Views.MergeWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Wukong_PBData_ReadWriter_GUI"
+        xmlns:local="clr-namespace:Wukong_PBData_ReadWriter_GUI.Views"
         mc:Ignorable="d"
         Title="MergeWindow" Height="430" Width="600">
     <Window.Resources>
